Standalone MLIR bytecode reader
Standalone, pure C, MLIR bytecode reader. This is WIP and represents a
reader structure that operates purely on handles and relies on callbacks
for customization. There is no single in-memory representation here,
instead the caller to specialize for given application. Conceptually
this is a generic visitor over the bytecode and/or an event-driven online
algorithm for parsing MLIR bytecode.

Currently this is incomplete and there is duplication with C++ code, but
goal is to push this to further state where that duplication can be
removed.

The tablegen helper is primarily a bootstrapping convenience.

Re-exported from dev branch.